% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anifood}
\alias{anifood}
\title{case-control data}
\format{
A data frame with 500 rows and 11 variables:
\describe{
  \item{case}{case status, 1 = case, 0 = control}
  \item{exp01}{whether exposed to exp01, 1 = yes, 0 = no}
  \item{exp09}{whether exposed to exp09, 1 = yes, 0 = no}
  \item{exp20}{whether exposed to exp20, 1 = yes, 0 = no}
  \item{exp24}{whether exposed to exp24, 1 = yes, 0 = no}
  \item{exp27}{whether exposed to exp27, 1 = yes, 0 = no}
  \item{exp43}{whether exposed to exp43, 1 = yes, 0 = no}
  \item{exp45}{whether exposed to exp45, 1 = yes, 0 = no}
  \item{exp50}{whether exposed to exp50, 1 = yes, 0 = no}
  \item{exp52}{whether exposed to exp52, 1 = yes, 0 = no}
  \item{exp57}{whether exposed to exp57, 1 = yes, 0 = no}
}
}
\usage{
anifood
}
\description{
A toy dataset containing 7-day exposure history of 250 cases and 250 controls
}
\keyword{datasets}
