% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excl_vars}
\alias{excl_vars}
\title{Variables excluded from matching}
\format{
A data frame with two variables:
\describe{
  \item{exp_var}{exposures of interest}
  \item{rm_vars}{variables to be excluded from matching for a given exposure}
}
}
\usage{
excl_vars
}
\description{
A dataset lists variables that are excluded from matching for each exposure.
This dataset is supplied to the \code{rmvars} argument of the function
\code{\link{make_knn_strata}}. The two columns must be named with "exp_var" and "rm_vars".
}
\keyword{datasets}
