% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_data.R
\name{finalize_data}
\alias{finalize_data}
\title{Final cleaning of the matched dataset(s)}
\usage{
finalize_data(dfs, filter = TRUE, filterdata = NULL)
}
\arguments{
\item{dfs}{A list of data frames generated by
\code{\link{make_analysis_sets}}}

\item{filter}{Filter statement to apply}

\item{filterdata}{Extra data to left join to the \code{dfs} for filtering}
}
\value{
A list of data frames
}
\description{
Ensures that a control retained in a data frame is used once and remove strata
without any case or any control. In this process, priority is first given
to the smallest strata then smallest distance if a control is matched to
multiple cases (i.e., that control exists in multiple strata).
}
\details{
For more information, please refer to the vignette using
\code{browseVignettes("nncc")}.
}
