% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_df.R
\name{fix_df}
\alias{fix_df}
\title{Fix the strata so they all have at least one case and control}
\usage{
fix_df(d)
}
\arguments{
\item{d}{A stratified dataset}
}
\description{
Fix the strata so they all have at least one case and control
}
