% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_analysis_set.R
\name{make_analysis_sets}
\alias{make_analysis_sets}
\title{Make analysis datasets}
\usage{
make_analysis_sets(stratified_data, expvars, data, threshold)
}
\arguments{
\item{stratified_data}{List of stratified data sets, see
\code{\link{make_knn_strata}}}

\item{expvars}{Character vector of exposure variable for each set in
\code{stratifed_data}}

\item{data}{Original case control data}

\item{threshold}{Maximum distance threshold for cases and controls created by
\code{\link{get_threshold}}}
}
\value{
A list of data frames with the \code{length} of number of exposures
}
\description{
This helper function facilitates the implement the make_analysis_set() to
each exposure.
}
\details{
For more information, please refer to the vignette using
  \code{browseVignettes("nncc")}.
}
