% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_threshold.R
\name{original_compare_plot}
\alias{original_compare_plot}
\title{Compare the original strata's distances to the knn version}
\usage{
original_compare_plot(data, casevar, stratavar, threshold_results)
}
\arguments{
\item{data}{The original data}

\item{casevar}{The variable that defines cases vs. controls}

\item{stratavar}{The variable that defines the strata}

\item{threshold_results}{See \code{\link{get_threshold}}}
}
\value{
An list with items:
   \item{plot_density}{The ggplot displayed}
   \item{prop_distance_gt_threshold}{A table showing proportion of pairs exceeding numeric threshold chosen}
}
\description{
Compare the original strata's distances to the knn version
}
