% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot the OR results}
\usage{
plot_results(csvfilename, filter = TRUE)
}
\arguments{
\item{csvfilename}{CSV results file, see \code{\link{write_strata_or_output}}}

\item{filter}{How to filter the results}
}
\value{
Returns \code{csvfilename} to allow chaining
}
\description{
Plot the OR results
}
\details{
For more information, please refer to the vignette using
  \code{browseVignettes("nncc")}.
}
