% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_threshold.R
\name{threshold_model_plot}
\alias{threshold_model_plot}
\title{Show the prediction of the logistic regression model}
\usage{
threshold_model_plot(threshold_results, p_threshold = 0.5)
}
\arguments{
\item{threshold_results}{See \code{\link{get_threshold}}}

\item{p_threshold}{The probability that the closest matching approach
produces the closer matching relative to the random matching approach.
The greater \code{p_threshold}, the smaller the threshold.}
}
\value{
The ggplot showing the threshold logistic regression model
}
\description{
Show the prediction of the logistic regression model
}
