% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_controls.R
\name{unique_controls}
\alias{unique_controls}
\title{Ensures controls are unique to avoid potential pseudoreplication issues}
\usage{
unique_controls(stratifieddata)
}
\arguments{
\item{stratifieddata}{See \code{\link{make_knn_strata}} and \code{\link{make_analysis_set}}.}
}
\description{
Ensures controls are unique to avoid potential pseudoreplication issues
}
