% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_strata_or.R
\name{write_strata_or_output}
\alias{write_strata_or_output}
\title{Format strata output into CSV}
\usage{
write_strata_or_output(results, varnames, filename)
}
\arguments{
\item{results}{Output of \code{\link{calc_strata_or}}}

\item{filename}{String of the filename to output to}

\item{varnams}{Vector of exposure variable names}
}
\value{
Returns the filename to allow chaining
}
\description{
Format strata output into CSV
}
