% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheit.R
\name{cacheit}
\alias{cacheit}
\title{Function to cache long operations}
\usage{
cacheit(name, code, dir = "cache", createdir = FALSE, clearcache = FALSE)
}
\arguments{
\item{name}{Name of the file to create}

\item{code}{Expression of the code to execute and cache}

\item{dir}{Name of directory for cache to be placed in the working directory}

\item{createdir}{Logical about whether to create the directory if it does not exist}

\item{clearcache}{Recalculate the cache for this object}
}
\value{
Output of code, either freshly executed if the file exists or from the cache file
}
\description{
Function to cache long operations
}
