% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_strata_or.R
\name{calc_strata_or}
\alias{calc_strata_or}
\title{Calculate the pooled strata ORh}
\usage{
calc_strata_or(dfs, filter = TRUE, filterdata = NULL)
}
\arguments{
\item{dfs}{A named list of dataframes created by package functions}

\item{filter}{Filter statement to apply}

\item{filterdata}{Extra data to left join to the \code{dfs} for filtering}
}
\description{
Calculate the pooled strata ORh
}
\details{
Uses the M-H method unless there is only one strata for which the fisher.test is used
}
