% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLIP.confint_difflevel.R
\name{CLIP.confint.difflevel}
\alias{CLIP.confint.difflevel}
\title{Combine confidence interval for \code{logistf::logistf()} after multiple imputation}
\usage{
CLIP.confint.difflevel(
  obj = NULL,
  variable = NULL,
  data,
  firth = TRUE,
  weightvar = NULL,
  control = logistf:::logistf.control(),
  ci.level = c(0.025, 0.975),
  pvalue = TRUE,
  offset = NULL,
  bound.lo = NULL,
  bound.up = NULL,
  legacy = FALSE
)
}
\arguments{
\item{obj}{Either a list of logistf fits (on multiple imputed data sets), or the result of analysis of a \code{mice} (multiply imputed) object using \code{with.mids}}

\item{variable}{Must be used to include variables of interest; each of
variable of interest must have the same levels across different imputed
data sets.}

\item{data}{A list of data set corresponding to the model fits. Can be left blank if obj was obtained with the \code{dataout=TRUE} option or if obj was obtained by mice}

\item{firth}{If \code{TRUE}, applies the Firth correction. Should correspond to the entry in obj.}

\item{weightvar}{An optional weighting variable for each observation.}

\item{control}{Control parameters for \code{logistf}, usually obtained by \code{logistf.control()}}

\item{ci.level}{The two confidence levels for each tail of the posterior distribution.}

\item{pvalue}{If \code{TRUE}, will also compute a P-value from the posterior.}

\item{offset}{An optional offset variable}

\item{bound.lo}{Bounds (vector of length 2) for the lower limit. Can be left blank. Use only if problems are encountered.}

\item{bound.up}{Bounds (vector of length 2) for the upper limit. Can be left blank. Use only if problems are encountered.}

\item{legacy}{If \code{TRUE}, will use pure R code for all model fitting. Can be slow. Not recommended.}
}
\description{
The function was modified from \code{\link[logistf]{CLIP.confint}} to combine
results from m imputed data sets that have different structures (e.g., a
covariate in a model have different levels across different imputed data
sets) on April 15, 2022.
}
\details{
The formula in \code{\link[logistf]{logistf}} must be written as variable of
interest followed by covariates that have different levels across different
imputed data sets.

For more information, please refer to the vignette using
\code{browseVignettes("nncc")} and the original function
\code{\link[logistf]{CLIP.confint}}.

Please cite the original function \code{\link[logistf]{CLIP.confint}} for
publication.
}
