% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_threshold.R
\name{distance_density_plot}
\alias{distance_density_plot}
\title{Distance density plots comparing closest to random choices}
\usage{
distance_density_plot(threshold_results)
}
\arguments{
\item{threshold_results}{See \code{\link{get_threshold}}}
}
\value{
The ggplot showing the distances of cases matched to their nearest neighbor vs. a random control
}
\description{
Distance density plots comparing closest to random choices
}
