% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_threshold.R
\name{get_threshold}
\alias{get_threshold}
\title{What is the right threshold}
\usage{
get_threshold(data, vars, case_var = "case", seed = 1600)
}
\arguments{
\item{data}{The dataset}

\item{vars}{The variables to use for calculating distance}

\item{use_cache}{Filename for \code{\link{cacheit}} or NULL if you do not want to cache the results}
}
\description{
What is the right threshold
}
\details{
This function uses the logistic regression to predict by the distance
whether a control is the closest (unique) match for each case vs.
a random selection and returns the 50% threshold
}
