% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_analysis_set.R
\name{make_analysis_set}
\alias{make_analysis_set}
\title{Make analysis set}
\usage{
make_analysis_set(stratified_data, var, data, maxdist = 0, maxcontrols = 20,
  silent = FALSE)
}
\arguments{
\item{data}{Original case control data}

\item{maxdist}{Reject any controls more than maxdist from their case}

\item{maxcontrols}{Maximum number of controls to keep per strata}

\item{silent}{Suppress exposure info useful for *apply/loop implementations?}

\item{stratifed_data}{Stratified dataset, see \code{\link{make_knn_strata}}}

\item{expvar}{Character of current exposure variable in d}
}
\description{
Allow each control to only appear once,
keeping the closest case-control pair
after merging exposure data and removing strata with 
cases who are missing exposure information completely
}
