% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_analysis_set.R
\name{make_analysis_sets}
\alias{make_analysis_sets}
\title{Make analysis datasets}
\usage{
make_analysis_sets(stratified_data, expvars, data, threshold)
}
\arguments{
\item{expvars}{Character vector of exposure variable for each set in \code{stratifed_data}}

\item{data}{Original case control data}

\item{threshold}{Maximum distance threshold for cases and controls}

\item{stratifed_data}{List of stratified datasets, see \code{\link{make_knn_strata}}}
}
\description{
make_analysis_sets(cc, strata2, cc_vars, threshhold_results$threshold)
}
