% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_knn_strata.R
\name{make_knn_strata}
\alias{make_knn_strata}
\title{Make case-control strata using k nearest neighbors (knn)}
\usage{
make_knn_strata(expvar, matchvars, df, casevar = "case", ncntls = 250,
  metric = "gower", silent = FALSE)
}
\arguments{
\item{expvar}{Character - name of exposure variable in df.}

\item{matchvars}{Character vector - what are the variables to match on.  Note that the function automatically excludes the
the exposure varible when placed in the list which facilitates *apply function and loop constructs}

\item{df}{Data.frame - data set.}

\item{casevar}{Charater - what is the variable indicating case status (1 = case, 0 = control)}

\item{ncntls}{Numeric - number of controls to find for each case (k in knn).}

\item{metric}{Character - metric to be used. See \code{\link{cluster::daisy}}.}

\item{silent}{Suppress exposure info useful for *apply/loop implementations?}
}
\description{
Make case-control strata using k nearest neighbors (knn)
}
