% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_threshold.R
\name{threshold_model_plot}
\alias{threshold_model_plot}
\title{Show the prediction of the logistic regression model}
\usage{
threshold_model_plot(threshold_results)
}
\arguments{
\item{threshold_results}{See \code{\link{get_threshold}}}
}
\description{
Show the prediction of the logistic regression model
}
