% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_controls.R
\name{unique_controls}
\alias{unique_controls}
\title{Ensures controls are unique to avoid possible pseudoreplication issues}
\usage{
unique_controls(stratifieddata)
}
\arguments{
\item{stratifieddata}{See \code{\link{make_knn_strata}} and
\code{\link{make_analysis_set}}.}
}
\value{
A tibble after it has been examined and filtered for duplicate controls
}
\description{
Ensures controls are unique to avoid possible pseudoreplication issues
}
